refactor(components): move components into nested directories
