+++ lib/components/ContentPlaceholder/ContentPlaceholder.vue
