+++ lib/components/Legend/OrdinalLegend.vue
