+++ lib/components/Legend/ScaleLegend.vue
