+++ lib/components/index.ts
@@ -1,36 +1,36 @@
-export { default as AccordionWrapper } from './AccordionWrapper.vue'
-export { default as AccordionStep } from './AccordionStep.vue'
-export { default as ActiveTextTruncate } from './ActiveTextTruncate.vue'
-export { default as AdvancedLinkForm } from './AdvancedLinkForm.vue'
-export { default as AdvancedLinkFormTab } from './AdvancedLinkFormTab.vue'
-export { default as ButtonIcon } from './ButtonIcon.vue'
-export { default as ButtonIconCounter } from './ButtonIconCounter.vue'
-export { default as BrandExpansion } from './BrandExpansion.vue'
-export { default as Brand } from './Brand.vue'
-export { default as ConfirmButton } from './ConfirmButton.vue'
-export { default as ContentPlaceholder } from './ContentPlaceholder.vue'
-export { default as CustomPagination } from './CustomPagination.vue'
-export { default as DigitsInput } from './DigitsInput.vue'
-export { default as DonateForm } from './DonateForm.vue'
-export { default as EmbeddableFooter } from './EmbeddableFooter.vue'
-export { default as EmbedForm } from './EmbedForm.vue'
-export { default as FollowUsPopover } from './FollowUsPopover.vue'
-export { default as GenericFooter } from './GenericFooter.vue'
-export { default as GenericHeader } from './GenericHeader.vue'
-export { default as HapticCopy } from './HapticCopy.vue'
+export { default as AccordionWrapper } from './Accordion/AccordionWrapper.vue'
+export { default as AccordionStep } from './Accordion/AccordionStep.vue'
+export { default as ActiveTextTruncate } from './ActiveTextTruncate/ActiveTextTruncate.vue'
+export { default as AdvancedLinkForm } from './AdvancedLinkForm/AdvancedLinkForm.vue'
+export { default as AdvancedLinkFormTab } from './AdvancedLinkForm/AdvancedLinkFormTab.vue'
+export { default as Brand } from './Brand/Brand.vue'
+export { default as BrandExpansion } from './Brand/BrandExpansion.vue'
+export { default as ButtonIcon } from './Button/ButtonIcon.vue'
+export { default as ButtonIconCounter } from './Button/ButtonIconCounter.vue'
+export { default as ConfirmButton } from './Button/ConfirmButton.vue'
+export { default as ContentPlaceholder } from './ContentPlaceholder/ContentPlaceholder.vue'
+export { default as CustomPagination } from './Pagination/CustomPagination.vue'
+export { default as DigitsInput } from './DigitsInput/DigitsInput.vue'
+export { default as DonateForm } from './DonateForm/DonateForm.vue'
+export { default as EmbeddableFooter } from './EmbeddableFooter/EmbeddableFooter.vue'
+export { default as EmbedForm } from './EmbedForm/EmbedForm.vue'
+export { default as FollowUsPopover } from './FollowUsPopover/FollowUsPopover.vue'
+export { default as GenericFooter } from './Generic/GenericFooter.vue'
+export { default as GenericHeader } from './Generic/GenericHeader.vue'
+export { default as HapticCopy } from './HapticCopy/HapticCopy.vue'
 export { default as ImageMode } from './ImageMode/ImageMode.vue'
 export { default as ImageModeSource } from './ImageMode/ImageModeSource.vue'
-export { default as OrdinalLegend } from './OrdinalLegend.vue'
-export { default as PhosphorIcon } from './PhosphorIcon.vue'
-export { default as PhosphorIconLayers } from './PhosphorIconLayers.vue'
-export { default as RangePicker } from './RangePicker.vue'
-export { default as ResponsiveIframe } from './ResponsiveIframe.vue'
-export { default as ScaleLegend } from './ScaleLegend.vue'
-export { default as SelectableDropdown } from './SelectableDropdown.vue'
-export { default as SecretInput } from './SecretInput.vue'
-export { default as SharingOptions } from './SharingOptions.vue'
-export { default as SharingOptionsLink } from './SharingOptionsLink.vue'
-export { default as SignUpForm } from './SignUpForm.vue'
-export { default as SlideUpDown } from './SlideUpDown.vue'
-export { default as TexturedDeck } from './TexturedDeck.vue'
-export { default as TinyPagination } from './TinyPagination.vue'
+export { default as OrdinalLegend } from './Legend/OrdinalLegend.vue'
+export { default as PhosphorIcon } from './PhosphorIcon/PhosphorIcon.vue'
+export { default as PhosphorIconLayers } from './PhosphorIcon/PhosphorIconLayers.vue'
+export { default as RangePicker } from './RangePicker/RangePicker.vue'
+export { default as ResponsiveIframe } from './ResponsiveIframe/ResponsiveIframe.vue'
+export { default as ScaleLegend } from './Legend/ScaleLegend.vue'
+export { default as SecretInput } from './SecretInput/SecretInput.vue'
+export { default as SelectableDropdown } from './SelectableDropdown/SelectableDropdown.vue'
+export { default as SharingOptions } from './SharingOptions/SharingOptions.vue'
+export { default as SharingOptionsLink } from './SharingOptions/SharingOptionsLink.vue'
+export { default as SignUpForm } from './SignUpForm/SignUpForm.vue'
+export { default as SlideUpDown } from './SlideUpDown/SlideUpDown.vue'
+export { default as TexturedDeck } from './TexturedDeck/TexturedDeck.vue'
+export { default as TinyPagination } from './Pagination/TinyPagination.vue'
